Integrate Input System
Replace input management from the legacy Unity's Input Manager
to the Unity's Input System package.

Assets:
- InputSystem_Actions.inputactions: Update the default input actions asset
removing unecessary actions and leaving just the actions "Move" and
"Attacking" that are relevant for the player controls (Opted to use the
default input action because it already contains all schemes configured
and the actions for UI-related actions).
- Game.unity: Add reference to the InputActionAsset into the Player GameObject
- ProjectSettings.asset: Disable legacy Input Manager

Scripts:
- `PlayerController`: Refactored the player controller logic using the
InputActionAsset in place of the legacy Input Manager usage.
- `Player`: Move the player movement logic from `PlayerController` to
this script.
