+++ Assets/Scripts/PlayerController.cs
@@ -1,58 +1,34 @@
 using UnityEngine;
+using UnityEngine.InputSystem;
 
 public class PlayerController : MonoBehaviour
 {
-    public float speed = 5;
-    public Vector2 Direction { get; private set; } = Vector2.zero;
+    [SerializeField] private InputActionAsset actions;
+    private InputAction actionMove;
 
-    private Collider2D coll;
+    public Vector2 Direction
+    {
+        get
+        {
+            if (!enabled)
+                return Vector2.zero;
+                
+            return actionMove.ReadValue<Vector2>().normalized;
+        }
+    }
 
     private void Start()
     {
-        coll = GetComponent<Collider2D>();
+        actionMove = actions["Move"];
+    }
+
+    private void OnEnable()
+    {
+        actions.Enable();
     }
 
     private void OnDisable()
     {
-        Direction = Vector2.zero;
-    }
-
-    private void Update()
-    {
-        ProcessInputs();
-        Move();
-        ClampPositionToScreen();
-    }
-
-    private void ProcessInputs()
-    {
-        Direction = new Vector2(
-            Input.GetAxisRaw("Horizontal"),
-            Input.GetAxisRaw("Vertical")
-        ).normalized;
-    }
-
-    private void Move()
-    {
-        transform.Translate(speed * Time.deltaTime * Direction);
-    }
-
-    private void ClampPositionToScreen()
-    {
-        Vector3 maxPosition = Camera.main.ScreenToWorldPoint(new Vector3(Screen.width, Screen.height, 0));
-        Vector3 minPosition = Camera.main.ScreenToWorldPoint(new Vector3(0, 0, 0));
-
-        Vector2 extents = coll.bounds.extents;
-
-        minPosition.x += extents.x;
-        minPosition.y += extents.y;
-
-        maxPosition.x -= extents.x;
-        maxPosition.y -= extents.y;
-
-        Vector3 position = transform.position;
-        position.x = Mathf.Clamp(position.x, minPosition.x, maxPosition.x);
-        position.y = Mathf.Clamp(position.y, minPosition.y, maxPosition.y);
-        transform.position = position;
+        actions.Disable();
     }
 }
